requirements.txt
spacy
gspread
oauth2client
google-api-python-client
google-auth
google-auth-oauthlib